refactor: move validation to shared layer
