+++ src/features/auth/ui/LoginForm.ts
@@ -1,7 +1,6 @@
 import { Router } from "../../../app/router/router";
 import { AuthState } from "../model/AuthState";
-import { Form, TextInputParams } from "../../../shared";
-import { ValidationPatterns } from "../model/validation";
+import { Form, TextInputParams, ValidationPatterns } from "../../../shared";
 
 const LOGIN_FORM_FIELDS: Array<TextInputParams> = [
   {
